Accept input from stdin
By popular convention, CLI applications that accept from stdin
designate the dash or hyphen character ('-') as the standard
input file descriptor. This commit adds specialized logic to treat
the '-' file path differently. When the user provides this character
in place of a file, the application will accept input from standard input.

On the off-chance the user would like to refer to an entry named `-`
in the current directory, they can specify a slightly more explicit
path: `./-`.